Faltando separar função de salvar para despesa e receita
