+++ src/components/DynamicForm/DynamicForm.module.css
@@ -14,6 +14,14 @@
     box-shadow: 0 0 10px rgba(0, 0, 0, 0.2);
 }
 
+.form-container.spentState {
+    background-color: #ffe5e5; /* Vermelho claro */
+  }
+  
+  .form-container.inputState {
+    background-color: #e5ffe5; /* Verde claro */
+  }
+
 /* Título do Formulário */
 .form-title {
     font-size: 1.5rem;
@@ -69,7 +77,7 @@
     border-radius: 0.375rem;
     padding: 0.5rem;
     cursor: pointer;
-    max-width: 20rem;
+    
     border: none;
     transition: background-color 0.3s ease;
 }
@@ -127,10 +135,12 @@
 @media (max-width: 600px) {
     .form-section {
         padding: 1rem;
+        
     }
 
     .form-input,
     .form-select {
         max-width: 100%;
+        min-width: 100%;
     }
 }
