+++ src/components/resumeStyles/homeResume.module.css
@@ -1,34 +1,51 @@
 .containerSection {
-    border: 1px solid var(--zinc-400);
-    border-radius: 0.375rem;
-    display: flex;
-    flex-direction: column;
-    gap: 0.5rem;
-    padding: 1rem;
-    min-width: 300px;
-    max-width: 100%;
-    background-color: white;
-    box-shadow: 0 0 10px rgba(0, 0, 0, 0.2);
-    margin: 2em;
-  }
-  .formContainer{
-    display: flex;
-    gap: 0.5rem;
-    justify-content: space-between;
-   
-    overflow-x: hidden; /* Esconde rolagem horizontal */
-    
-    box-sizing: border-box; /* Evita que padding aumente a largura */
-  }
-  .form{
-    max-width: 45%; /* Limita a largura máxima de cada formulário */
-  }
-  .resumeTitle {
-    margin: 0;
-    padding: 1rem 3rem 0.5rem 3rem;
-    color: var(--zinc-500);
-    font-size: 1.2rem;
+  border: 1px solid var(--zinc-400);
+  border-radius: 0.375rem;
+  display: flex;
+  flex-direction: column;
+  gap: 0.5rem;
+  padding: 1rem;
+  min-width: 300px;
+  max-width: 100%;
+  background-color: white;
+  box-shadow: 0 0 10px rgba(0, 0, 0, 0.2);
+  margin: 2em;
+  box-sizing: border-box; /* Certifica que o padding não vai fazer o container ultrapassar a largura */
 }
+
+.formContainer {
+  display: flex;
+  gap: 0.5rem;
+  justify-content: space-between;
+  overflow-x: hidden; /* Esconde rolagem horizontal */
+  box-sizing: border-box; /* Inclui padding e bordas no cálculo de largura */
+  flex-wrap: wrap; /* Permite que os itens se movam para a linha seguinte */
+}
+
+.form {
+  width: 100%; /* Faz com que os formulários ocupem toda a largura disponível */
+  max-width: 45%; /* Limita a largura máxima de cada formulário */
+  box-sizing: border-box; /* Inclui padding e bordas no cálculo da largura */
+}
+
+.resumeTitle {
+  margin: 0;
+  padding: 1rem 3rem 0.5rem 3rem;
+  color: var(--zinc-500);
+  font-size: 1.2rem;
+}
+
 .secundaryTitle {
-    font-size: x-large;
+  font-size: x-large;
+}
+
+/* Media query para telas pequenas */
+@media (max-width: 768px) {
+  .formContainer {
+      flex-direction: column; /* Empilha os elementos em uma coluna em telas pequenas */
   }
+
+  .form {
+      max-width: 100%; /* Deixa os formulários com largura de 100% em telas pequenas */
+  }
+}
