+++ src/styles/homeResume.module.css
@@ -1,0 +1,35 @@
+.containerSection {
+  border: 1px solid var(--zinc-400);
+  border-radius: 0.375rem;
+  display: flex;
+  flex-direction: column;
+  gap: 0.5rem;
+  padding: 1rem;
+  min-width: 300px;
+  max-width: 100%;
+  background-color: white;
+  box-shadow: 0 0 10px rgba(0, 0, 0, 0.2);
+  margin: 2em;
+}
+.formContainer{
+  display: flex;
+  gap: 0.5rem;
+  justify-content: center;
+  
+ 
+  overflow-x: hidden; /* Esconde rolagem horizontal */
+  
+  box-sizing: border-box; /* Evita que padding aumente a largura */
+}
+.form{
+  max-width: 45%; /* Limita a largura máxima de cada formulário */
+}
+.resumeTitle {
+  margin: 0;
+  padding: 1rem 3rem 0.5rem 3rem;
+  color: var(--zinc-500);
+  font-size: 1.2rem;
+}
+.secundaryTitle {
+  font-size: x-large;
+}
